feat: 이름변경 및 memberQuery추가
